pyproject.toml
[tool.poetry]
name = "just-do"
version = "0.1.0"
description = ""
authors = ["Anatoly Boronilov <boronilov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
djoser = "^2.2.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
django-environ = "^0.11.2"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
